feat: Create a v2 of get_outputs_from_cloudformation_stack
I was getting bogged down in the weeds trying to make the original script work for both `appName` and `service` and decided to call it quits.

This v2 version can be used by projects that have updated to use the field `service` instead of `appName` in `cdk.json`. Love it or leave it. :muscle: